+++ src/inference/tests/functional/ov_extension_test.hpp
@@ -1,0 +1,157 @@
+// Copyright (C) 2018-2023 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+//
+#pragma once
+
+#include "common_test_utils/test_common.hpp"
+#include "openvino/runtime/core.hpp"
+
+class OVExtensionTests : public CommonTestUtils::TestsCommon {
+public:
+    ov::Core core;
+
+    void test() {
+        std::string model = R"V0G0N(
+<net name="Activation" version="10">
+    <layers>
+        <layer name="in1" type="Parameter" id="0" version="opset1">
+            <data shape="1,3,22,22" element_type="f32"/>
+            <output>
+                <port id="0" precision="FP32" names="in_data">
+                    <dim>1</dim>
+                    <dim>3</dim>
+                    <dim>22</dim>
+                    <dim>22</dim>
+                </port>
+            </output>
+        </layer>
+        <layer name="activation" id="1" type="Identity" version="extension">
+            <input>
+                <port id="1" precision="FP32">
+                    <dim>1</dim>
+                    <dim>3</dim>
+                    <dim>22</dim>
+                    <dim>22</dim>
+                </port>
+            </input>
+            <output>
+                <port id="2" precision="FP32" names="out_data">
+                    <dim>1</dim>
+                    <dim>3</dim>
+                    <dim>22</dim>
+                    <dim>22</dim>
+                </port>
+            </output>
+        </layer>
+        <layer name="output" type="Result" id="2" version="opset1">
+            <input>
+                <port id="0" precision="FP32">
+                    <dim>1</dim>
+                    <dim>3</dim>
+                    <dim>22</dim>
+                    <dim>22</dim>
+                </port>
+            </input>
+        </layer>
+    </layers>
+    <edges>
+        <edge from-layer="0" from-port="0" to-layer="1" to-port="1"/>
+        <edge from-layer="1" from-port="2" to-layer="2" to-port="0"/>
+    </edges>
+</net>
+)V0G0N";
+        ov::Tensor weights;
+        ov::PartialShape refBeforeReshape{1, 3, 22, 22};
+        ov::PartialShape refAfterReshape{8, 9, 33, 66};
+
+        auto network = core.read_model(model, weights);
+        std::map<std::string, ov::PartialShape> newShapes;
+        newShapes["in_data"] = refAfterReshape;
+
+        EXPECT_EQ(refBeforeReshape, network->output().get_partial_shape());
+        EXPECT_NO_THROW(network->reshape(newShapes));
+        EXPECT_EQ(refAfterReshape, network->output().get_partial_shape());
+    }
+
+    void test_two_op() {
+        std::string model = R"V0G0N(
+<net name="Activation" version="10">
+    <layers>
+        <layer name="in1" type="Parameter" id="0" version="opset1">
+            <data shape="1,3,22,22" element_type="f32"/>
+            <output>
+                <port id="0" precision="FP32" names="in_data">
+                    <dim>1</dim>
+                    <dim>3</dim>
+                    <dim>22</dim>
+                    <dim>22</dim>
+                </port>
+            </output>
+        </layer>
+        <layer name="activation" id="1" type="Identity" version="extension">
+            <input>
+                <port id="1" precision="FP32">
+                    <dim>1</dim>
+                    <dim>3</dim>
+                    <dim>22</dim>
+                    <dim>22</dim>
+                </port>
+            </input>
+            <output>
+                <port id="2" precision="FP32" names="out_data">
+                    <dim>1</dim>
+                    <dim>3</dim>
+                    <dim>22</dim>
+                    <dim>22</dim>
+                </port>
+            </output>
+        </layer>
+        <layer name="activation2" id="2" type="CustomReLU" version="extension">
+            <input>
+                <port id="1" precision="FP32">
+                    <dim>1</dim>
+                    <dim>3</dim>
+                    <dim>22</dim>
+                    <dim>22</dim>
+                </port>
+            </input>
+            <output>
+                <port id="2" precision="FP32" names="out_relu_data">
+                    <dim>1</dim>
+                    <dim>3</dim>
+                    <dim>22</dim>
+                    <dim>22</dim>
+                </port>
+            </output>
+        </layer>
+        <layer name="output" type="Result" id="3" version="opset1">
+            <input>
+                <port id="0" precision="FP32">
+                    <dim>1</dim>
+                    <dim>3</dim>
+                    <dim>22</dim>
+                    <dim>22</dim>
+                </port>
+            </input>
+        </layer>
+    </layers>
+    <edges>
+        <edge from-layer="0" from-port="0" to-layer="1" to-port="1"/>
+        <edge from-layer="1" from-port="2" to-layer="2" to-port="1"/>
+        <edge from-layer="2" from-port="2" to-layer="3" to-port="0"/>
+    </edges>
+</net>
+)V0G0N";
+        ov::Tensor weights;
+        ov::PartialShape refBeforeReshape{1, 3, 22, 22};
+        ov::PartialShape refAfterReshape{8, 9, 33, 66};
+
+        auto network = core.read_model(model, weights);
+        std::map<std::string, ov::PartialShape> newShapes;
+        newShapes["in_data"] = refAfterReshape;
+
+        EXPECT_EQ(refBeforeReshape, network->output().get_partial_shape());
+        EXPECT_NO_THROW(network->reshape(newShapes));
+        EXPECT_EQ(refAfterReshape, network->output().get_partial_shape());
+    }
+};
